Refactor elasticjob-tracing-spring-boot-starter (#1191)
* Update elasticjob-tracing-spring-boot-starter pom

* Update javadoc

* Update pom of elasticjob-registry-center-spring-boot-starter

* Rename package boot.job
