+++ elasticjob-lite/elasticjob-lite-spring/elasticjob-lite-spring-boot-starter/src/test/resources/META-INF/services/org.apache.shardingsphere.elasticjob.executor.item.JobItemExecutor
@@ -14,5 +14,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-org.apache.shardingsphere.elasticjob.lite.spring.boot.executor.CustomClassedJobExecutor
-org.apache.shardingsphere.elasticjob.lite.spring.boot.executor.PrintJobExecutor
+org.apache.shardingsphere.elasticjob.lite.spring.boot.job.executor.CustomClassedJobExecutor
+org.apache.shardingsphere.elasticjob.lite.spring.boot.job.executor.PrintJobExecutor
